Ambient Light Levels (#196)
For "Allow designers to specify inside vs outside levels" task. 

Adds two new variables, use_fixed_ambient and fixed_ambient_color.
Use_fixed_ambient selects from a list of premade colors or a custom
color. If it is set on level load, it overwrites the day night cycle and
disables the color change from the sun (Day night cycle still runs in
the background, just cant change the color of the screen). If using the
custom color, fixed_ambient_color provides the color. Uses an enum for
the selectable options and a light dictionary (_interior_light) for the
premade color filters.

---------

Co-authored-by: Sunwolf <sunspotforest@gmail.com>