feat: allow the panel to be placed globally
This updates the panel so that it can be placed globally in the world,
this works by adding a presentation canvas layer to the world. We can't
directly add to the world or the tween of the day/night cycle will tween
everything. So a new canvas layer is added that follows the viewport.