feat: add interact panel for interaction presentation (#197)
This adds the UI/UX and plumbing need to present interactions to a
player. This adds the interact panel along with its components such as
menus and options. It also adds support for the examine interaction
which has its own key binding and presentation. Due to this requirement
the panel and friends has been designed in as generic a fashion as
possible so that is can be used for all our interaction needs.


Fixes #119