Adds style outline (#222)
Creates a custom Dialogic style. 
It is currently bare-bones but sets up the needed resources to implement
changes to the dialog set up in the future. Restyles the choice buttons
and makes them compatible with the Rich Text format, allowing for text
color changes, bolding, italics, and strikethroughs to be added and
removed dynamically during gameplay. Implements choice history and
strikes through choices that have been previously selected.

---------

Co-authored-by: Sunwolf <sunspotforest@gmail.com>
Co-authored-by: this-is-envy <this-is-envy@users.noreply.github.com>