Multi-action toasts work; NPCs can hook into them (#227)
So Many Things.

- Rewrite how state transitions work. In adding multi-action events that
involved multiple dialogic timelines we had to await during event
execution which caused _problems_. See the writeup at the end for
details.
- Revisit multi-action panel & sub-panel selection
- Adds 4 configurable actions for NPCs
- Adds a second configurable action for Greybox objects
- Tweak some audio notifications to reduce wildly verbose logging

**State transfer issue writeup**

<details>

Here's the explainer I did initially. I think it's mostly correct but I
never did 100% nail down the cause vs "likely" cause of the issue I was
solving for.

> in order to sort out one last issue with the multi-input toast thing I
ended up running into the "can move while in dialogue" bug and i would
advise avoiding that one the fix is going to be gnarly -- basically
what's happening is that because we trigger the dialogue from an Effect
which isn't in an Interact state we don't ignore inputs (allowing
movement)
>
> So to fix that you basically need to convert the Effect application to
coroutines using await effect.act(...)
>
> But doing that is viral and pushes itself up the stack so that other
stuff works bc now the functions calling act aren't blocking unless you
await on them etc
>
> So you do that up the chain until everything is okay except now you
don't exit the Interact state as you expect. Which is hard because the
problem there is that await makes the change_state call asynchronous
which means it's no longer assured to happen in a run_$step section and
you essentially throw away the state change because each run_<step>
starts with _next_step = null.
>
> I don't remember exactly why we needed to do that but I'm hesitant to
just change that and am not sure what the Right solution is rn

</details>

---------

Co-authored-by: this-is-envy <this-is-envy@users.noreply.github.com>